MSVC: Enable checksum in Release binaries
The openvpn-build/windows-msi expects binaries to have a valid checksum
in their PE header. While MinGW build provides one, the MSVC does not
(by default).

This allows packaging of the MSVC-built binaries with
openvpn-build/windows-msi.

The checksum generation was enabled only in Release builds. It conflicts
with "Edit and Continue" feature of MSVC used in Debug builds.

Signed-off-by: Simon Rozman <simon@rozman.si>